+++ lib/chat_web/live/main_live/page/feed.ex
@@ -1,6 +1,6 @@
 defmodule ChatWeb.MainLive.Page.Feed do
   @moduledoc "Lobby part of chat. User list and room list"
-  import Phoenix.Component
+  use ChatWeb, :live_component
 
   alias Chat.Log
   alias Chat.User
@@ -12,8 +12,17 @@ defmodule ChatWeb.MainLive.Page.Feed do
 
     socket
     |> assign(:action_feed_till, till)
-    |> assign(:action_feed_list, list)
+    |> assign(:items, nil)
+    |> assign_feed_stream(list)
     |> assign(:feed_update_mode, :ignore)
+  end
+
+  def render(assigns) do
+    ~H"""
+    <%= for item <- @action_feed_list do %>
+      <.item item={item} tz={@tz} />
+    <% end %>
+    """
   end
 
   def more(%{assigns: %{action_feed_till: since}} = socket) do
@@ -21,17 +30,23 @@ defmodule ChatWeb.MainLive.Page.Feed do
 
     socket
     |> assign(:action_feed_till, till)
-    |> assign(:action_feed_list, list)
+    |> assign_feed_stream(list)
     |> assign(:feed_update_mode, :append)
   end
 
   def close(socket) do
     socket
     |> assign(:action_feed_till, nil)
-    |> assign(:action_feed_list, nil)
+    |> clean_feed_items()
+    |> assign(:items, nil)
   end
 
-  def item(%{item: {who, data}, tz: timezone} = assigns) do
+  def item(
+        %{
+          item: {dom_id, {_uuid, who, data}},
+          tz: timezone
+        } = assigns
+      ) do
     datetime =
       data
       |> elem(0)
@@ -43,13 +58,16 @@ defmodule ChatWeb.MainLive.Page.Feed do
       assign(assigns,
         user: User.by_id(who),
         action: data |> Tuple.delete_at(0),
-        datetime: datetime
+        datetime: datetime,
+        dom_id: dom_id
       )
 
     ~H"""
-    <div class="border-0 rounded-md bg-white/20 p-2 flex flex-col justify-start">
-      <span class="text-white"><%= @user && @user.name %> <.action action={@action} /></span>
-      <div class="text-white/70" style="font-size: 10px;"><%= @datetime %></div>
+    <div class="py-1 flex justify-start" id={@dom_id}>
+      <div class="border-0 rounded-md bg-white/20 p-2 flex flex-col justify-start">
+        <span class="text-white"><%= @user && @user.name %> <.action action={@action} /></span>
+        <div class="text-white/70" style="font-size: 10px;"><%= @datetime %></div>
+      </div>
     </div>
     """
   end
@@ -123,5 +141,35 @@ defmodule ChatWeb.MainLive.Page.Feed do
     else
       load_more(rest_count, small_list ++ list, till - 1)
     end
+  end
+
+  defp assign_feed_stream(%{assigns: %{streams: %{action_feed_list: _feed}}} = socket, list) do
+    socket
+    |> stream_batch_insert(:action_feed_list, list,
+      at: -1,
+      dom_id: &"item-#{elem(&1, 0)}"
+    )
+    |> assign_items_uuid(list)
+  end
+
+  defp assign_feed_stream(socket, list) do
+    socket
+    |> stream(:action_feed_list, list, dom_id: &"item-#{elem(&1, 0)}")
+    |> assign_items_uuid(list)
+  end
+
+  defp assign_items_uuid(%{assigns: %{items: nil}} = socket, list) do
+    socket |> assign(:items, Enum.map(list, &"item-#{elem(&1, 0)}"))
+  end
+
+  defp assign_items_uuid(%{assigns: %{items: items}} = socket, list) do
+    socket |> assign(:items, Enum.map(list, &"item-#{elem(&1, 0)}") ++ items)
+  end
+
+  defp clean_feed_items(socket) do
+    socket.assigns.items
+    |> Enum.reduce(socket, fn item, socket ->
+      stream_delete_by_dom_id(socket, :action_feed_list, item)
+    end)
   end
 end